pyproject.toml
[tool.poetry]
name = "test_parser"
version = "1.0"
description = ""
authors = ["mavrinia <mavrinia@er.gov35.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.4"
celery = "^5.4.0"
python-dotenv = "^1.0.1"
django-webpack-loader = "^3.1.0"
isort = "^5.13.2"
black = "^24.4.2"
flake8 = "^7.0.0"
radon = "^6.0.1"
xenon = "^0.9.1"
djangorestframework-datatables = "^0.7.1"
djangorestframework = "^3.15.1"
django-select2 = "^8.1.2"
django-redis = "^5.4.0"
django-bootstrap5 = "^24.2"
requests = "^2.31.0"
django-filter = "^24.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
